Rename Script works, edited tab index in several forms.
Rename Script drop down menu item populates a new form.
Allows user to select the script they want to rename, grabs the value
from the textbox and renames the script name in the XML.

Currently there is no functionality to edit the location of the script -
Users will need to remove the script and re-add the script to do that.
Or manually modify the XML file.

Adjusted the tab index in the forms to be logical. Enabled KeyPreview
and coded proper keyletters/acceptbutton/cancelbuttons for all forms
except the main form.